feat: make directory data extensible
This commit makes it possible to add custom data to the directory type.
Additionally, it creates a new entry type which is, itself, extensible.
Clients may store custom data in either of these types. Because they are
included in the directory data, these types are part of the Merkle tree
and gain all the properties therein.

Further, we can carry this extensibility all the way to the root
directory if tags are modified to take an entry type rather than simply
a tree hash.

One use case for this functionality is when a client wants to mirror
data from one tree item into another tree or even sub-tree. Under the
previous data structure this was possible, but with one important
drawback: the fact that the mirrored tree was mirrored and the origin of
where it was mirrored from were both lost in the process. Retaining this
additional metadata is important for forensic reconstruction.

Under the new data type, however, this data can be retained. For
example, if one were to mirror python:3.10.0 as a subtree in a new tree
at the location `/python` along with some application code, the
directory data for root could look like this:

```
{
  "application.py": { "hash": "sha256:..." },
  "python": {
    "hash": "sha256:...",
    "from": "python:3.10.0"
  }
}
```

Signed-off-by: Nathaniel McCallum <nathaniel@profian.com>