feat(api): include metadata in entry
Instead of storing a digest of exherent properties of a node, store
the exherent properties directly, which simplifies API and
implementations.

Note, that since a content digest is required on API level for tag and
tree node creation, integrity of the data is ensured with the same
guarantees as original approach.

With this change, the server may return additional digests, which were
not specified on upload on fetch to further improve security guarantees.

Signed-off-by: Roman Volosatovs <roman@profian.com>