feat: authenticate users using OIDC
The Drawbridge server now authenticates users using the `Bearer` token
header in requests. The token is validated with the OIDC provider, which
is discovered at startup.
Each OIDC subject is associated with at most one user.
OIDC subjects are store on-disk within an `oidc-label` namespace for
future flexibility and possiblity of migration.

Responsibility of creating a child entity was moved into the parent in
`store` crate. That is necessary for e.g. the user creation, because the
store has to not only create and populate the user entity directory, but
also create a symlink of form `/oidc/<label>/<subj>` -> `/users/<name>`

Signed-off-by: Roman Volosatovs <roman@profian.com>