feat: check access token
Change verification of Access Token to actually verify JWT signatures
and audience/scope claims.
This prevents tokens issued for one service or environment to be used
here, and for tokens acquired for limited permissions to be used for
others.

It also means no network requests need to be performed during
verification of requests.

Signed-off-by: Patrick Uiterwijk <patrick@profian.com>