refactor: do not `rewind` files in the tree
`rewind` is redundant as implementations actually using
internally-stored file contents can do it themselves.

By avoiding the `rewind`, the read position is now also communicated to
the caller, which is the correct behavior - for example, in case a file
was written to after the tree construction, if `rewind` was to be used, it would
not be possible to determine, which file content was hashed without validating the hash.

Signed-off-by: Roman Volosatovs <roman@profian.com>